+++ vitest.config.js
@@ -22,8 +22,8 @@ export default defineConfig({
       thresholds: {
         lines: 80,
         functions: 80,
-        statements: 75,
-        branches: 60
+        statements: 80,
+        branches: 80
       }
     }
   }